# openapi.yaml
...
{
  "openapi": "3.1.0",
  "info": {
    "title": "Афиша рядом API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Карточка события",
        "operationId": "get_event_api_v1_events__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90,
                  "minimum": -90
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180,
                  "minimum": -180
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "events"
        ],
        "summary": "Изменить событие",
        "operationId": "patch_event_api_v1_events__event_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventManage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Удалить черновик",
        "operationId": "delete_event_api_v1_events__event_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/queue": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Очередь модерации",
        "operationId": "queue_api_v1_admin_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/events/{event_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Карточка события для модератора",
        "operationId": "event_card_api_v1_admin_events__event_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventCard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/events/{event_id}/decision": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Решение по событию",
        "operationId": "decide_event_api_v1_admin_events__event_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDecisionIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/verifications/{request_id}/decision": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Решение по заявке на проверку",
        "operationId": "decide_verification_api_v1_admin_verifications__request_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orgs/{org_id}/revoke": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Отозвать верификацию",
        "operationId": "revoke_api_v1_admin_orgs__org_id__revoke_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Журнал аудита",
        "operationId": "audit_api_v1_admin_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "before_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminEventCard": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventManage"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminAuthor"
              },
              {
                "type": "null"
              }
            ]
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/RuleFlag"
            },
            "type": "array",
            "title": "Flags"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionItem"
            },
            "type": "array",
            "title": "Decisions"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/AuditItem"
            },
            "type": "array",
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "event",
          "author",
          "flags",
          "decisions",
          "history"
        ],
        "title": "AdminEventCard",
        "description": "Карточка модератора: событие, автор, флаги правил, решения и журнал (новые сверху)."
      },
...
      "EventDecisionIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "reject",
              "hide"
            ],
            "title": "Action"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "EventDecisionIn"
      },
      "EventDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "trust_tier": {
            "type": "string",
            "title": "Trust Tier"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "org": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrgBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "venue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VenueBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "locality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalityBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "description": "Часовой пояс населённого пункта события"
          },
          "next_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "sessions_count": {
            "type": "integer",
            "title": "Sessions Count"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "price_type": {
            "type": "string",
            "title": "Price Type"
          },
          "price_min": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Min"
          },
          "price_max": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max"
          },
          "pushkin_card": {
            "type": "boolean",
            "title": "Pushkin Card"
          },
          "age_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Rating"
          },
          "is_online": {
            "type": "boolean",
            "title": "Is Online"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/SessionOut"
            },
            "type": "array",
            "title": "Sessions"
          },
          "map_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Map Url"
          },
          "online_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Online Url"
          },
          "ticket_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Url"
          },
          "registration_required": {
            "type": "boolean",
            "title": "Registration Required"
          },
          "contacts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contacts"
          },
          "accessibility": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accessibility"
          },
          "indoor": {
            "type": "string",
            "title": "Indoor"
          },
          "source": {
            "$ref": "#/components/schemas/SourceInfo"
          },
          "ai_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ai Fields"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_past": {
            "type": "boolean",
            "title": "Is Past"
          },
          "saved_session_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Saved Session Ids"
          },
          "share_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Share Url",
            "description": "Диплинк https://max.ru/<бот>?startapp=ev_<id> для шеринга"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "short_description",
          "category",
          "cover_url",
          "trust_tier",
          "is_demo",
          "org",
          "venue",
          "locality",
          "timezone",
          "next_session",
          "sessions_count",
          "distance_km",
          "price_type",
          "price_min",
          "price_max",
          "pushkin_card",
          "age_rating",
          "is_online",
          "description",
          "tags",
          "sessions",
          "map_url",
          "online_url",
          "ticket_url",
          "registration_required",
          "contacts",
          "accessibility",
          "indoor",
          "source",
          "ai_fields",
          "status",
          "is_past"
        ],
        "title": "EventDetail"
      },
...
      "EventPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "short_description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 32,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "cover_media_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Media Id"
          },
          "venue_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Id"
          },
          "locality_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locality Id"
          },
          "is_online": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Online"
          },
          "online_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Online Url"
          },
          "indoor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Indoor"
              },
              {
                "type": "null"
              }
            ]
          },
          "price_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "price_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Min"
          },
          "price_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max"
          },
          "pushkin_card": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pushkin Card"
          },
          "age_rating": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  0,
                  6,
                  12,
                  16,
                  18
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Rating"
          },
          "registration_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Required"
          },
          "ticket_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Url"
          },
          "contacts": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Contacts"
          },
          "accessibility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Accessibility"
              },
              {
                "type": "null"
              }
            ]
          },
          "sessions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SessionIn"
                },
                "type": "array",
                "maxItems": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EventPatch",
        "description": "Меняются только переданные поля; `sessions` заменяет список целиком."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueueOut": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/QueueEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "verifications": {
            "items": {
              "$ref": "#/components/schemas/QueueVerification"
            },
            "type": "array",
            "title": "Verifications"
          }
        },
        "type": "object",
        "required": [
          "events",
          "verifications"
        ],
        "title": "QueueOut"
      },
...
      "RevokeIn": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RevokeIn"
      },
...
      "VerificationDecisionIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Action"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "VerificationDecisionIn"
      },
      "VerificationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "org_id": {
            "type": "integer",
            "title": "Org Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "Разместить на странице организации"
          },
          "site_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Url"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/VerificationStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At",
            "description": "Раньше этого повтор недоступен"
          }
        },
        "type": "object",
        "required": [
          "id",
          "org_id",
          "method",
          "status",
          "code",
          "site_url",
          "steps",
          "decision_reason",
          "created_at",
          "last_attempt_at",
          "next_attempt_at"
        ],
        "title": "VerificationOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT из POST /api/v1/auth/max или /auth/guest",
        "scheme": "bearer"
      }
    }
  }
}